refactor: mode loadIframe as helper
