+++ packages/metascraper-audio/package.json
@@ -17,9 +17,7 @@
   ],
   "dependencies": {
     "@metascraper/helpers": "^5.16.16",
-    "async-memoize-one": "~1.0.1",
-    "cheerio": "~1.0.0-rc.5",
-    "jsdom": "~16.4.0"
+    "cheerio": "~1.0.0-rc.5"
   },
   "devDependencies": {
     "mocha": "latest",
