+++ packages/metascraper-helpers/package.json
@@ -27,6 +27,7 @@
     "is-uri": "~1.2.0",
     "iso-639-3": "~2.2.0",
     "isostring": "0.0.1",
+    "jsdom": "~16.4.0",
     "lodash": "~4.17.20",
     "memoize-one": "~5.1.1",
     "microsoft-capitalize": "~1.0.5",
