+++ packages/metascraper-telegram/package.json
@@ -20,7 +20,6 @@
     "@metascraper/helpers": "^5.16.16",
     "async-memoize-one": "~1.0.1",
     "css-urls": "~2.0.11",
-    "jsdom": "~16.4.0",
     "tldts": "~5.7.2"
   },
   "devDependencies": {
